requirements.txt
h5py==2.8.0
hdijupyterutils==0.12.5
html5lib==1.0.1
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
nltk==3.3
notebook==5.5.0
numpy==1.14.5
pandas==0.22.0
prompt-toolkit==1.0.15
protobuf==3.5.2
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.7.3
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.1.0
torch==0.4.0
torchvision==0.2.1
tqdm==4.23.4


